Updated documentation to reflect the better-looking method syntax requirements. Fixed issues with store class. Will start implementations.
