+++ libs/core/include/oxide/core/buffer.h
@@ -1,5 +1,5 @@
 #pragma once
 
-#include "buffer/ibuffer.h"
+#include "buffer/i_buffer.h"
 #include "buffer/static_buffer.h"
 #include "buffer/buffer_factory.h"
