+++ libs/core/include/oxide/core/buffer/static_buffer.h
@@ -1,5 +1,5 @@
 #pragma once
-#include "ibuffer.h"
+#include "i_buffer.h"
 #include <array>
 #include <cstring>
 #include <oxide/core/export.h>
@@ -31,7 +31,7 @@ public:
         return {};
     }
 
-    std::expected<void, Error> shrinkToFit() override {
+    std::expected<void, Error> shrink_to_fit() override {
         // No-op for static buffer
         return {};
     }
