+++ libs/core/include/oxide/core/time/clock.h
@@ -6,9 +6,9 @@
 
 namespace oxide::core {
 
-class OXIDE_CORE_API ITimer {
+class OXIDE_CORE_API IClock {
 public:
-    virtual ~ITimer() = default;
+    virtual ~IClock() = default;
     
     [[nodiscard]]
     virtual uint64_t now_ns() const = 0;
